Update the prettier option to wrap markdown files. The license bolierplate is manually merged into a single line to avoid errors when generating the htmls by Docusaurus.
When we do git commit, eslint and prettier will run according to
the configuration files in docs-site/ folder.

The end of line setting in prettier is to always convert crlf to lf.
There is no difference in display in moder editors and browsers.
In some windows system, if the eol is lf, then two lines will show
in one line; In some linux system, if the eol is crlf, then there
will be an extra ^M at the start (or end) of the line.